Better error message around misnaming of federation directive
Adds additional suggestions to "unknown directive" error messages when
either 1) the name is probably a typo or 2) we recognize a default
federation directive name, but it is the wrong name for that directive
in the schema due either the directive not being imported, or being
imported under a different name.

As an aside, this commit also update error handling in the `buildSchema`
method to avoid throwing on the first encountered error, but instead
collect all errors and throw a compound error at the end (as this is
both more convenient and more consistent with how `Schema.validate`
already works).