Improves validation of @link (mostly for `import`)
The type of the `@link(import")` is `[link__Import]`, where
`link__Import` is a scalar for which we accept either a string, or a
{name: "", as: ""}-like input object value). But we weren't properly
validating that concrete values had one of those 2 representations
and passing random values resulted in unhelpful errors. This is
fixed by this commit, which now return meaningful errors.

Additionally, another easy error to do is to try to import a name
that the feature actually doesn't "expose" (which could be a typo
or could be someone trying to import `"key"` instead of `"@key"`).
This commit also ensure that, for feature we actually know (one day,
we might be figure out which elements a feature exposes by following
its url, but currently we're limited to what we know), we reject
imports of unkown name with helpful messages.